Compress kart control action

+++ src/karts/controller/player_controller.hpp
@@ -29,7 +29,7 @@ class PlayerController : public Controller
 friend class KartRewinder;
 protected:
     int            m_steer_val, m_steer_val_l, m_steer_val_r;
-    int            m_prev_accel;
+    uint16_t       m_prev_accel;
     bool           m_prev_brake;
     bool           m_prev_nitro;
